rasp-pi/requirements.txt
mysql-connector-python==8.3.0
enviroplus
pms5003
bme280